Moving scatter and gather functions to a new package
