+++ biopet-extensions/src/main/scala/nl/lumc/sasc/biopet/extensions/gatk/scatter/LocusScatterFunction.scala
@@ -12,7 +12,7 @@
  * license; For commercial users or users who do not want to follow the AGPL
  * license, please contact us to obtain a separate license.
  */
-package nl.lumc.sasc.biopet.extensions.gatk
+package nl.lumc.sasc.biopet.extensions.gatk.scatter
 
 import org.broadinstitute.gatk.queue.function.InProcessFunction
 import org.broadinstitute.gatk.utils.interval.IntervalUtils
@@ -20,19 +20,14 @@ import org.broadinstitute.gatk.utils.interval.IntervalUtils
 import scala.collection.JavaConversions._
 
 /**
- * Splits intervals by contig instead of evenly.
+ * A scatter function that divides down to the locus level.
  */
-class ContigScatterFunction extends GATKScatterFunction with InProcessFunction {
-
-  override def scatterCount = if (intervalFilesExist) super.scatterCount min this.maxIntervals else super.scatterCount
-
-  protected override def maxIntervals = {
-    GATKScatterFunction.getGATKIntervals(this.originalGATK).contigs.size
-  }
+class LocusScatterFunction extends GATKScatterFunction with InProcessFunction {
+  protected override def maxIntervals = scatterCount
 
   def run() {
     val gi = GATKScatterFunction.getGATKIntervals(this.originalGATK)
-    IntervalUtils.scatterContigIntervals(gi.samFileHeader, gi.locs, this.scatterOutputFiles)
+    val splits = IntervalUtils.splitLocusIntervals(gi.locs, this.scatterOutputFiles.size)
+    IntervalUtils.scatterFixedIntervals(gi.samFileHeader, splits, this.scatterOutputFiles)
   }
 }
-
